port/engine: clang-tidy pass over the C++ side
Bug-tier: three GetPointerSize() implementations returned
sizeof(std::vector) instead of the buffer's byte size, two AddRectFilled
calls passed their rounding/flags arguments swapped, FRotator's
IRotator conversion multiplied by 360/UINT16_MAX (integer zero), and
five polymorphic base classes gained virtual destructors.

Cleanup: override specifiers tree-wide, duplicate includes dropped,
reserved-identifier header guards and one member renamed, overflow-prone
size multiplications widened, a compound assignment split out of an if.

+++ src/engine/sky/SkyCloud.cpp
@@ -96,7 +96,7 @@ SkyCloud::SkyCloud(ScreenContext* screen, u16 cloudVariant, u16 posY, u16 rotY, 
 // foreign pixels (e.g. the minimap) once the allocator reuses the memory.
 void SkyCloud::ResolveTexture() {
     if (GameEngine_ResourceGetTexTypeByName((const char*)CM_GetProps()->CloudTexture) != 1) { // Stock
-        mTexture = ((u8*) LOAD_ASSET_RAW(CM_GetProps()->CloudTexture)) + (mCloudVariant * 1024);
+        mTexture = ((u8*) LOAD_ASSET_RAW(CM_GetProps()->CloudTexture)) + ((size_t) mCloudVariant * 1024);
         mVtx = (Vtx*)D_0D005FB0;
     } else { // Texture pack: pass the asset name, resolved fresh at import time
         mTexture = CM_GetProps()->CloudTexture;
